chore: Update web tests
