--- packages/article-main-standard/src/article-prop-types/article-prop-types.base.js
@@ -1,21 +1,0 @@
-import PropTypes from "prop-types";
-import ArticleHeader from "../article-header/article-header";
-import ArticleMeta from "../article-meta/article-meta";
-
-const articlePropTypesBase = {
-  adConfig: PropTypes.shape({}).isRequired,
-  analyticsStream: PropTypes.func.isRequired,
-  data: PropTypes.shape({
-    ...ArticleHeader.propTypes,
-    ...ArticleMeta.propTypes
-    // Could have more here
-  }),
-  receiveChildList: PropTypes.func
-};
-
-const articleDefaultPropsBase = {
-  data: null,
-  receiveChildList: () => {}
-};
-
-export { articlePropTypesBase, articleDefaultPropsBase };
